revamp payments
