+++ app/src/main/java/com/socialite/solite_pos/data/source/local/entity/room/new_master/Payment.kt
@@ -6,6 +6,7 @@ import androidx.room.Index
 import androidx.room.PrimaryKey
 import androidx.sqlite.db.SimpleSQLiteQuery
 import com.socialite.solite_pos.data.source.local.room.AppDatabase.Companion.UPLOAD
+import com.socialite.solite_pos.data.source.remote.response.entity.PaymentResponse
 import com.socialite.solite_pos.view.ui.DropdownItem
 import java.io.Serializable
 import java.util.UUID
@@ -46,6 +47,18 @@ data class Payment(
     fun asNewPayment() = this.copy(
         id = UUID.randomUUID().toString()
     )
+
+    fun toResponse(): PaymentResponse {
+        return PaymentResponse(
+                id = id,
+                name = name,
+                desc = desc,
+                tax = tax,
+                isCash = isCash,
+                isActive = isActive,
+                isUploaded = true
+        )
+    }
 
     companion object {
         const val ID = "id_payment"
